fuzz test working

+++ src/MerklePatricia.sol
@@ -36,11 +36,11 @@ library MerklePatricia {
      * @param keys a list of keys to verify
      * @return bytes[] a list of values corresponding to the supplied keys.
      */
-    function VerifySubstrateProof(
-        bytes32 root,
-        bytes[] memory proof,
-        bytes[] memory keys
-    ) internal pure returns (StorageValue[] memory) {
+    function VerifySubstrateProof(bytes32 root, bytes[] memory proof, bytes[] memory keys)
+        internal
+        pure
+        returns (StorageValue[] memory)
+    {
         StorageValue[] memory values = new StorageValue[](keys.length);
         TrieNode[] memory nodes = new TrieNode[](proof.length);
 
@@ -51,9 +51,7 @@ library MerklePatricia {
         for (uint256 i = 0; i < keys.length; i++) {
             values[i].key = keys[i];
             NibbleSlice memory keyNibbles = NibbleSlice(keys[i], 0);
-            NodeKind memory node = SubstrateTrieDB.decodeNodeKind(
-                TrieDB.get(nodes, root)
-            );
+            NodeKind memory node = SubstrateTrieDB.decodeNodeKind(TrieDB.get(nodes, root));
 
             // worst case scenario, so we avoid unbounded loops
             for (uint256 j = 0; j < MAX_TRIE_DEPTH; j++) {
@@ -66,35 +64,22 @@ library MerklePatricia {
                     }
                     break;
                 } else if (TrieDB.isNibbledBranch(node)) {
-                    NibbledBranch memory nibbled = SubstrateTrieDB
-                        .decodeNibbledBranch(node);
-                    uint256 nibbledBranchKeyLength = NibbleSliceOps.len(
-                        nibbled.key
-                    );
+                    NibbledBranch memory nibbled = SubstrateTrieDB.decodeNibbledBranch(node);
+                    uint256 nibbledBranchKeyLength = NibbleSliceOps.len(nibbled.key);
                     if (!NibbleSliceOps.startsWith(keyNibbles, nibbled.key)) {
                         break;
                     }
 
-                    if (
-                        NibbleSliceOps.len(keyNibbles) == nibbledBranchKeyLength
-                    ) {
+                    if (NibbleSliceOps.len(keyNibbles) == nibbledBranchKeyLength) {
                         if (Option.isSome(nibbled.value)) {
                             values[i].value = TrieDB.load(nodes, nibbled.value.value);
                         }
                         break;
                     } else {
-                        uint256 index = NibbleSliceOps.at(
-                            keyNibbles,
-                            nibbledBranchKeyLength
-                        );
-                        NodeHandleOption memory handle = nibbled.children[
-                            index
-                        ];
+                        uint256 index = NibbleSliceOps.at(keyNibbles, nibbledBranchKeyLength);
+                        NodeHandleOption memory handle = nibbled.children[index];
                         if (Option.isSome(handle)) {
-                            keyNibbles = NibbleSliceOps.mid(
-                                keyNibbles,
-                                nibbledBranchKeyLength + 1
-                            );
+                            keyNibbles = NibbleSliceOps.mid(keyNibbles, nibbledBranchKeyLength + 1);
                             nextNode = handle.value;
                         } else {
                             break;
@@ -104,9 +89,7 @@ library MerklePatricia {
                     break;
                 }
 
-                node = SubstrateTrieDB.decodeNodeKind(
-                    TrieDB.load(nodes, nextNode)
-                );
+                node = SubstrateTrieDB.decodeNodeKind(TrieDB.load(nodes, nextNode));
             }
         }
 
@@ -122,12 +105,11 @@ library MerklePatricia {
      * @param childInfo data that can be used to compute the root of the child trie
      * @return bytes[], a list of values corresponding to the supplied keys.
      */
-    function ReadChildProofCheck(
-        bytes32 root,
-        bytes[] memory proof,
-        bytes[] memory keys,
-        bytes memory childInfo
-    ) internal pure returns (StorageValue[] memory) {
+    function ReadChildProofCheck(bytes32 root, bytes[] memory proof, bytes[] memory keys, bytes memory childInfo)
+        internal
+        pure
+        returns (StorageValue[] memory)
+    {
         // fetch the child trie root hash;
         bytes memory prefix = bytes(":child_storage:default:");
         bytes memory key = bytes.concat(prefix, childInfo);
@@ -148,11 +130,11 @@ library MerklePatricia {
      * @param keys a list of keys to verify
      * @return bytes[] a list of values corresponding to the supplied keys.
      */
-    function VerifyEthereumProof(
-        bytes32 root,
-        bytes[] memory proof,
-        bytes[] memory keys
-    ) external pure returns (StorageValue[] memory) {
+    function VerifyEthereumProof(bytes32 root, bytes[] memory proof, bytes[] memory keys)
+        external
+        pure
+        returns (StorageValue[] memory)
+    {
         StorageValue[] memory values = new StorageValue[](keys.length);
         TrieNode[] memory nodes = new TrieNode[](proof.length);
 
@@ -163,9 +145,7 @@ library MerklePatricia {
         for (uint256 i = 0; i < keys.length; i++) {
             values[i].key = keys[i];
             NibbleSlice memory keyNibbles = NibbleSlice(keys[i], 0);
-            NodeKind memory node = EthereumTrieDB.decodeNodeKind(
-                TrieDB.get(nodes, root)
-            );
+            NodeKind memory node = EthereumTrieDB.decodeNodeKind(TrieDB.get(nodes, root));
 
             // worst case scenario, so we avoid unbounded loops
             for (uint256 j = 0; j < MAX_TRIE_DEPTH; j++) {
@@ -174,30 +154,19 @@ library MerklePatricia {
                 if (TrieDB.isLeaf(node)) {
                     Leaf memory leaf = EthereumTrieDB.decodeLeaf(node);
                     // Let's retrieve the offset to be used
-                    uint offset = keyNibbles.offset % 2 == 0
-                        ? keyNibbles.offset / 2
-                        : keyNibbles.offset / 2 + 1;
+                    uint256 offset = keyNibbles.offset % 2 == 0 ? keyNibbles.offset / 2 : keyNibbles.offset / 2 + 1;
                     // Let's cut the key passed as input
-                    keyNibbles = NibbleSlice(
-                        NibbleSliceOps.bytesSlice(keyNibbles.data, offset),
-                        0
-                    );
+                    keyNibbles = NibbleSlice(NibbleSliceOps.bytesSlice(keyNibbles.data, offset), 0);
                     if (NibbleSliceOps.eq(leaf.key, keyNibbles)) {
                         values[i].value = TrieDB.load(nodes, leaf.value);
                     }
                     break;
                 } else if (TrieDB.isExtension(node)) {
-                    Extension memory extension = EthereumTrieDB.decodeExtension(
-                        node
-                    );
+                    Extension memory extension = EthereumTrieDB.decodeExtension(node);
                     if (NibbleSliceOps.startsWith(keyNibbles, extension.key)) {
                         // Let's cut the key passed as input
                         keyNibbles = NibbleSlice(
-                            NibbleSliceOps.bytesSlice(
-                                keyNibbles.data,
-                                NibbleSliceOps.len(extension.key)
-                            ),
-                            0
+                            NibbleSliceOps.bytesSlice(keyNibbles.data, NibbleSliceOps.len(extension.key)), 0
                         );
                         nextNode = extension.node;
                     } else {
@@ -211,9 +180,7 @@ library MerklePatricia {
                         }
                         break;
                     } else {
-                        NodeHandleOption memory handle = branch.children[
-                            NibbleSliceOps.at(keyNibbles, 0)
-                        ];
+                        NodeHandleOption memory handle = branch.children[NibbleSliceOps.at(keyNibbles, 0)];
                         if (Option.isSome(handle)) {
                             keyNibbles = NibbleSliceOps.mid(keyNibbles, 1);
                             nextNode = handle.value;
@@ -225,9 +192,7 @@ library MerklePatricia {
                     break;
                 }
 
-                node = EthereumTrieDB.decodeNodeKind(
-                    TrieDB.load(nodes, nextNode)
-                );
+                node = EthereumTrieDB.decodeNodeKind(TrieDB.load(nodes, nextNode));
             }
         }
